:sparkles: Implement file manager

+++ file/blog_id.go
@@ -2,10 +2,10 @@ package file
 
 type BlockID struct {
 	fileName string
-	blockNum int
+	blockNum int64
 }
 
-func NewBlockID(name string, num int) *BlockID {
+func NewBlockID(name string, num int64) *BlockID {
 	return &BlockID{
 		fileName: name,
 		blockNum: num,
@@ -16,6 +16,6 @@ func (bid *BlockID) FileName() string {
 	return bid.fileName
 }
 
-func (bid *BlockID) Number() int {
+func (bid *BlockID) Number() int64 {
 	return bid.blockNum
 }
